requirements.txt
Flask>=3.0.0
psutil>=5.9.6
requests>=2.31.0
python-telegram-bot>=20.5
gunicorn>=21.2.0
werkzeug
pymongo
dnspython
